Resolved import issues and restructured components
